Use serial input for wideband AFR instead of analog.
I was having problems maintaining accurate readings over the analog output of the Innovate Motorsports LC-2 wideband, so
this change switches to using the serial output of the LC-2. The LC-2 outputs 0-5V RS-232 serial, so I'm using a voltage
divider (2.2K & 3.3K) to step down to 0-3V, and using RXINV since TTL and RS-232 serial are inverted.